test: Ready unit tests for testing multiple components.
