+++ test/unit/meson.build
@@ -5,6 +5,7 @@ cmocka_deps = dependency('cmocka', version: '>=1.1.5',
 
 libui_unit_sources = [
 	'main.c',
+	'slider.c',
 ]
 
 if libui_OS == 'windows'
